// src/domain/enums/campos-filtro-itens.ts
export enum CamposFiltroItens {
    areaConhecimentoFiltro = 'areaConhecimentoFiltro',
    disciplinaFiltro = 'disciplinaFiltro',
    anoMatrizFiltro = 'anoMatrizFiltro',
    categoriaItemFiltro = 'categoriaItemFiltro',
    dificuldadeSugeridaFiltro = 'dificuldadeSugeridaFiltro',
    situacaoItemFiltro = 'situacaoItemFiltro',
    matrizFiltro = 'matrizFiltro',
    competenciaFiltro = 'competenciaFiltro',
    habilidadeFiltro = 'habilidadeFiltro',
    informacoesEstatisticasFiltro = 'informacoesEstatisticasFiltro',
    palavraChaveFiltro = 'palavraChaveFiltro',
}